Enable bullet piercing when `obj_perice` is purchased

+++ objects/obj_enemy_walker/Collision_obj_bullet.gml
@@ -1,9 +1,15 @@
-instance_destroy(other); // kill the bullet
+var bullet_pierces = variable_global_exists("bullet_pierce") && global.bullet_pierce;
+
+if (!bullet_pierces) {
+    instance_destroy(other); // kill the bullet
+}
+
 instance_destroy();      // kill the enemy
 
 // Return shooting ability
-with (obj_player) {
-    can_shoot = true;
-    bullet_id = noone;
+if (!bullet_pierces) {
+    with (obj_player) {
+        can_shoot = true;
+        bullet_id = noone;
+    }
 }
-
